Externalize code regarding json validation from CardProcessingService (#3733)
